@@ -31,20 +31,20 @@
     "@types/jest": "^26.0.8",
     "@types/node": "^12",
     "@tsconfig/node12": "^1.0.7",
-    "@vercel/ncc": "^0.27.0",
+    "@vercel/ncc": "^0.33.4",
     "codecov": "^3.6.5",
     "eslint": "^7.25.0",
     "eslint-config-universe": "^5.0.0",
     "expo-yarn-workspaces": "^1.2.1",
     "husky": "^1.1.3",
-    "jest": "^26.2.2",
-    "jest-watch-typeahead": "^0.6.0",
+    "jest": "^27.5.1",
+    "jest-watch-typeahead": "^1.0.0",
     "lerna": "3.22.1",
     "lint-staged": "^8.0.4",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
     "memfs": "^2.15.5",
-    "typescript": "^4.2.4",
+    "typescript": "4.6.3",
     "yarn-deduplicate": "^3.1.0"
   },
   "resolutions": {
