@@ -1,7 +1,11 @@
-const path = require('path');
+import { dirname, resolve } from 'path';
+import { fileURLToPath } from 'url';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
-module.exports = {
+export default {
   preset: '../../jest/unit-test-config',
-  rootDir: path.resolve(__dirname),
-  displayName: require('./package').name,
+  rootDir: resolve(__dirname),
+  displayName: 'expo-env-info',
+  extensionsToTreatAsEsm: ['.ts'],
 };
